+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.2.1](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/compare/v1.2.0...v1.2.1) (2026-05-15)
+
+
+### Bug Fixes
+
+* fix issues ([35bb4b5](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/35bb4b56fea0ab63beb7d7e8cbfe7e95b97d51d8))
+* Fix Pipeline Permissions ([3700413](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/3700413fabc62e7bb3c71c72bdfadd20185fefc9))
+* Fix Pipeline Permissions ([2f0bf0f](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/2f0bf0fc2594dadbcc62a31d34ba6a9636a2f0ed))
+* fix pom accordingly with dependency check output ([b76882f](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/b76882f2441b98a2d14fe18d88243b0d6f6f6e0a))
+* fix pom accordingly with dependency check output ([5755873](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/5755873e7b6ba7099524c01fe3937d61a520b322))
+* fixes accordingly with new OWASP ZAP run ([8ccca57](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/8ccca57626b4d267eb5e658191f9dba952f1e333))
+* handle malformed urls ([08c16e2](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/08c16e22ae27f07da45c87d1c4c6fa4c2a6f554e))
+* implement tomcat config ([b712566](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/b712566d55dbe229d0dbb99caaa73af7b22abcd7))
+* protect MovieController with RoleGuard ([80d425e](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/80d425eef1a157226f211a31bacff0a56ae418af))
+* remove /errrors ([f316535](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/f3165350d492bd36103b007e9f25d7f01473951f))
+* remove some unneeded code ([b3e2ab7](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/b3e2ab7815f6315e2c669b539ed597ec56519b7a))
+* use jakarta in Movie domain class ([1587ee2](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/commit/1587ee29e18ed6920d000f1fcd99dfbb9d9356a1))
+
 ## [1.2.0](https://github.com/mei-desofs/desofs2026-wed_nap_1-1/compare/v1.1.1...v1.2.0) (2026-05-14)